Slice parent class out of EditOrderAddressFormViewModel
